+++ l2/eigen_service/src/app.ts
@@ -1,16 +1,18 @@
 import express from "express";
 import jwt from "express-jwt";
 import { v4 as uuidv4 } from "uuid";
+import cors from "cors";
+const TOTP = require("totp.js");
+import url from "url";
+
 import * as log4js from "./log";
 import * as db_pk from "./database_pk";
 import * as db_txh from "./database_transaction_history";
 import * as db_recovery from "./database_recovery";
+import * as friend_list from "./database_friend_relationship";
 import * as util from "./util";
-import url from "url";
 
 import { JWT_SECRET, TOTP_SECRET } from "./login/config";
-
-import cors from "cors";
 
 import * as userdb from "./pid/pid";
 
@@ -259,8 +261,368 @@ app.put("/txh/:txid", async function (req, res) {
   res.json(util.Succ(result));
 });
 
+// get user, his/her friends, his/her strangers by id
+app.get(
+  "/user/:user_id",
+  // jwt({ secret: JWT_SECRET, algorithms: ["HS256"] }),
+  async function (req, res) {
+    const user_id = req.params.user_id;
+    if (user_id === undefined) {
+      res.json(util.Err(-1, "invalid argument"));
+      return;
+    }
+    const action = req.query.action;
+
+    if (action === undefined) {
+      const result = await userdb.findByID(user_id);
+      console.log(result);
+      res.json(util.Succ(result));
+      return;
+    }
+
+    switch (action) {
+      case "guardians":
+        var filter_status = req.query.status;
+        if (filter_status !== undefined) {
+          console.log("Filter the status of guardians: ", filter_status);
+        }
+        if (user_id === undefined) {
+          //res.json(util.Err(-1, "invalid argument"));
+          var all_relationships = await friend_list.findAll();
+          res.json(util.Succ(all_relationships));
+          return;
+        }
+        if (!(await userdb.findByID(user_id))) {
+          console.log("The user does not exist ", user_id);
+          res.json(util.Err(-1, "user does not exist"));
+          return;
+        }
+        var status = await friend_list.getStatusByUserId(user_id);
+        var ids = new Set();
+        var relationships = new Map();
+        for (let i = 0; i < status.length; i++) {
+          // There isn't status filter or filter the status
+          if (
+            filter_status === undefined ||
+            filter_status == status[i].status
+          ) {
+            ids.add(status[i].user_id);
+            relationships[status[i].user_id] = status[i].status;
+          }
+        }
+        console.log(status, ids);
+        var information_without_status: any = await userdb.findUsersInformation(
+          Array.from(ids)
+        );
+        console.log("Infomation without status: ", information_without_status);
+        var information_with_status = new Array();
+        for (let i = 0; i < information_without_status.length; i++) {
+          information_with_status.push({
+            user_id: information_without_status[i].user_id,
+            email: information_without_status[i].email,
+            name: information_without_status[i].name,
+            status: relationships[information_without_status[i].user_id],
+          });
+        }
+        console.log(`Guardian list of ${user_id}: `, information_with_status);
+        res.json(util.Succ(information_with_status));
+        return;
+      case "strangers":
+        if (user_id === undefined) {
+          res.json(util.Err(-1, "invalid argument"));
+          return;
+        }
+        if (!(await userdb.findByID(user_id))) {
+          console.log("The user does not exist ", user_id);
+          res.json(util.Err(-1, "user does not exist"));
+          return;
+        }
+        var ids = await userdb.findAllUserIDs();
+        var known = await friend_list.getKnownByUserId(user_id);
+        var strangers = new Set([...ids].filter((x) => !known.has(x)));
+        strangers.delete(Number(user_id));
+        var result = Array.from(strangers);
+        var information = await userdb.findUsersInformation(result);
+
+        console.log(`Stranger list of ${user_id}: `, information);
+        res.json(util.Succ(information));
+        return;
+      default:
+        res.json(util.Err(-1, "invalid action"));
+        return;
+    }
+  }
+);
+
+// TODO: Just for test
+app.post(
+  "/user",
+  // jwt({ secret: JWT_SECRET, algorithms: ["HS256"] }),
+  async function (req, res) {
+    var result: any = await userdb.add(req.body);
+    console.log("Create a new user, id = ", result.user_id);
+    console.log(result);
+    return res.json(util.Succ(result));
+  }
+);
+
+// Guardian add
+app.post(
+  "/user/:user_id/guardian",
+  // jwt({ secret: JWT_SECRET, algorithms: ["HS256"] }),
+  async function (req, res) {
+    const user_id = req.params.user_id;
+    var guardian_id = req.body.guardian_id;
+    if (!util.has_value(user_id)) {
+      res.json(util.Err(-1, "missing user_id"));
+      return;
+    }
+
+    console.log(`User ${user_id} wants add guardian`);
+    const guardian_email = req.body.guardian_email;
+
+    if (guardian_id !== undefined && guardian_email) {
+      res.json(
+        util.Err(
+          -1,
+          "guardian_id and guardian_email can not exist at the same time"
+        )
+      );
+      return;
+    }
+
+    if (util.has_value(guardian_email)) {
+      var guardian = await userdb.findByEmail(guardian_email);
+      if (guardian) {
+        guardian_id = guardian.user_id;
+      } else {
+        res.json(util.Err(-1, "guardian_email do not exist in the database"));
+        return;
+      }
+    }
+
+    if (!util.has_value(guardian_id)) {
+      res.json(util.Err(-1, "miss guardian_id or guardian_email is not found"));
+      return;
+    }
+
+    if (
+      !(await userdb.findByID(user_id)) ||
+      !(await userdb.findByID(guardian_id))
+    ) {
+      console.log("One of the users does not exist", user_id, guardian_id);
+      res.json(util.Err(-1, "one of the users does not exist"));
+      return;
+    }
+
+    // NOTE: When send a friend requet, self is requester, guardian is responder
+    const result = await friend_list.request(user_id, guardian_id);
+    if (result) {
+      console.log("Send guardian request success!");
+      return res.json(util.Succ(result));
+    } else {
+      console.log("Send a guardian request fail!");
+      return res.json(util.Err(-1, "fail to send a guardian request"));
+    }
+  }
+);
+
+// Guardian confirm or reject
+app.put(
+  "/user/:user_id/guardian",
+  // jwt({ secret: JWT_SECRET, algorithms: ["HS256"] }),
+  async function (req, res) {
+    const user_id = req.params.user_id;
+    const action = req.body.action;
+    var guardian_id = req.body.guardian_id;
+    if (!util.has_value(user_id) || !util.has_value(action)) {
+      res.json(util.Err(-1, "missing user_id or action"));
+      return;
+    }
+
+    console.log(`User ${user_id} wants do ${action}`);
+    const guardian_email = req.body.guardian_email;
+
+    console.log(
+      `${action} is going to do: ${user_id}, ${guardian_id} or ${guardian_email}`
+    );
+
+    if (guardian_id !== undefined && guardian_email) {
+      res.json(
+        util.Err(
+          -1,
+          "guardian_id and guardian_email can not exist at the same time"
+        )
+      );
+      return;
+    }
+
+    if (util.has_value(guardian_email)) {
+      var guardian = await userdb.findByEmail(guardian_email);
+      if (guardian) {
+        guardian_id = guardian.user_id;
+      } else {
+        res.json(util.Err(-1, "guardian_email do not exist in the database"));
+        return;
+      }
+    }
+
+    if (!util.has_value(guardian_id)) {
+      res.json(util.Err(-1, "miss guardian_id or guardian_email is not found"));
+      return;
+    }
+
+    if (
+      !(await userdb.findByID(user_id)) ||
+      !(await userdb.findByID(guardian_id))
+    ) {
+      console.log("One of the users does not exist", user_id, guardian_id);
+      res.json(util.Err(-1, "one of the users does not exist"));
+      return;
+    }
+
+    var result;
+    switch (action) {
+      case "confirm":
+        // NOTE: When send a guardian confirm, self is responder, guardian is requester
+        result = await friend_list.confirm(guardian_id, user_id);
+        if (result) {
+          console.log("Confirm a guardian request success!");
+          return res.json(util.Succ(result));
+        } else {
+          console.log("Confirm a guardian request fail!");
+          return res.json(util.Err(-1, "fail to confirm a guardian request"));
+        }
+      case "reject":
+        // NOTE: When send a guardian reject, self is responder, guardian is requester
+        result = await friend_list.reject(guardian_id, user_id);
+        if (result) {
+          console.log("Reject a guardian request success!");
+          return res.json(util.Succ(result));
+        } else {
+          console.log("Reject a guardian fail!");
+          return res.json(util.Err(-1, "fail to reject a guardian request"));
+        }
+      default:
+        res.json(util.Err(-1, "invalid action"));
+        return;
+    }
+  }
+);
+
+// Guardian add
+app.delete(
+  "/user/:user_id/guardian",
+  // jwt({ secret: JWT_SECRET, algorithms: ["HS256"] }),
+  async function (req, res) {
+    const user_id = req.params.user_id;
+    var guardian_id = req.body.guardian_id;
+    if (!util.has_value(user_id)) {
+      res.json(util.Err(-1, "missing user_id"));
+      return;
+    }
+
+    console.log(`User ${user_id} wants delete guardian`);
+    const guardian_email = req.body.guardian_email;
+
+    if (guardian_id !== undefined && guardian_email) {
+      res.json(
+        util.Err(
+          -1,
+          "guardian_id and guardian_email can not exist at the same time"
+        )
+      );
+      return;
+    }
+
+    if (util.has_value(guardian_email)) {
+      var guardian = await userdb.findByEmail(guardian_email);
+      if (guardian) {
+        guardian_id = guardian.user_id;
+      } else {
+        res.json(util.Err(-1, "guardian_email do not exist in the database"));
+        return;
+      }
+    }
+
+    if (!util.has_value(guardian_id)) {
+      res.json(util.Err(-1, "miss guardian_id or guardian_email is not found"));
+      return;
+    }
+
+    if (
+      !(await userdb.findByID(user_id)) ||
+      !(await userdb.findByID(guardian_id))
+    ) {
+      console.log("One of the users does not exist", user_id, guardian_id);
+      res.json(util.Err(-1, "one of the users does not exist"));
+      return;
+    }
+
+    const result = await friend_list.remove(user_id, guardian_id);
+    if (result) {
+      console.log("Remove a guardian request success!");
+      return res.json(util.Succ(result));
+    } else {
+      console.log("Remove a guardian fail!");
+      return res.json(util.Err(-1, "fail to remove a guardian"));
+    }
+  }
+);
+
+app.put(
+  "/user/:user_id/otpauth",
+  // jwt({ secret: JWT_SECRET, algorithms: ["HS256"] },
+  async function (req, res) {
+    const user_id = req.params.user_id;
+    const secret = req.body.secret;
+    if (!util.has_value(user_id) || !util.has_value(secret)) {
+      res.json(util.Err(-1, "missing user_id or secret"));
+      return;
+    }
+
+    const result = await userdb.updateSecret(user_id, secret);
+
+    if (result) {
+      console.log("Save a otpauth secret success!");
+      res.json(util.Succ(result));
+      return;
+    } else {
+      console.log("Save a otpauth secret fail!");
+      res.json(util.Err(-1, "fail to save a otpauth secret"));
+      return;
+    }
+  }
+);
+
+// verify code
+app.post("/user/:user_id/otpauth", async function (req, res) {
+  const user_id = req.params.user_id;
+  const code = req.body.code;
+  if (!util.has_value(user_id) || !util.has_value(code)) {
+    return res.json(util.Err(1, "missing fields"));
+  }
+  console.log(req.body);
+  const user = await userdb.findByID(user_id);
+  if (user) {
+    if (user.secret) {
+      const totp = new TOTP(user.secret);
+      var result = totp.verify(code);
+      res.json(util.Succ(result));
+      return;
+    } else {
+      console.log("The secret does not exist ", user_id);
+      res.json(util.Err(-1, "secret does not exist"));
+      return;
+    }
+  } else {
+    console.log("The user does not exist ", user_id);
+    res.json(util.Err(-1, "user does not exist"));
+    return;
+  }
+});
+
 require("./login/google")(app);
-require("./pid/pid")(app);
 
 app.listen(3000, function () {
   console.log("Eigen Service listening on port 3000!");
